+++ handler/category_handler.go
@@ -11,8 +11,10 @@ import (
 
 func RegisterCategoryRoutes(e *echo.Echo) {
 	v1Group := e.Group("/v1")
-	locationGroup := v1Group.Group("/categories", middleware.AuthMiddleware)
-	locationGroup.GET("", getCategories)
+	catGroup := v1Group.Group("/categories", middleware.AuthMiddleware)
+	catGroup.GET("", getCategories)
+
+	catGroup.GET("/:id/list-product", getListProductByCategory)
 }
 
 func getCategories(c echo.Context) error {
@@ -22,4 +24,18 @@ func getCategories(c echo.Context) error {
 	}
 
 	return utils.ResponseJSON(c, "Success get data", data, http.StatusOK)
+}
+
+func getListProductByCategory(c echo.Context) error {
+	catID := utils.StrToUint64(c.Param("id"), 0)
+	if catID == 0 {
+		return utils.ResponseJSON(c, "Invalid category ID", nil, http.StatusBadRequest)
+	}
+
+	data, statusCode, err := repository.GetListProductByCategory(catID)
+	if err != nil {
+		return utils.ResponseJSON(c, err.Error(), nil, statusCode)
+	}
+
+	return utils.ResponseJSON(c, "Success", data, statusCode)
 }