+++ main.go
@@ -17,7 +17,7 @@ import (
 )
 
 const (
-	version = "1.0.6" // release-arvigo-backend-1.0.6
+	version = "1.0.7" // release-arvigo-backend-1.0.7
 	appName = "arvigo-backend"
 )