+++ repository/brand_repository.go
@@ -2,6 +2,7 @@ package repository
 
 import (
 	"net/http"
+	"strings"
 	"time"
 
 	"github.com/yusufwib/arvigo-backend/datastruct"
@@ -79,6 +80,31 @@ func UpdateBrand(brandID uint64, data datastruct.BrandInput) (statusCode int, er
 
 	if err = db.Where("id", brandID).Updates(&brandPayload).Error; err != nil {
 		return http.StatusInternalServerError, err
+	}
+
+	return
+}
+
+func GetListProductByBrand(brandID uint64) (res []datastruct.HomeProduct, statusCode int, err error) {
+	db := Database()
+	statusCode = http.StatusOK
+
+	if err := db.Table("products p").
+		Select([]string{
+			"p.id",
+			"p.name",
+			"p.images",
+			"b.name as brand",
+		}).
+		Joins("LEFT JOIN brands b on b.id = p.brand_id").
+		Where("p.merchant_id = 0 AND b.id = ?", brandID).
+		Find(&res).
+		Error; err != nil {
+		return res, http.StatusInternalServerError, err
+	}
+
+	for i, v := range res {
+		res[i].Image = strings.Split(v.Image, ",")[0]
 	}
 
 	return